chore: rename to emote
