+++ src/lib/babylon/emote.ts
