+++ src/lib/babylon/render.ts
@@ -2,7 +2,7 @@ import { Color4 } from '@babylonjs/core'
 import { PreviewConfig, PreviewType, BodyShape } from '@dcl/schemas'
 import { getBodyShape } from './body'
 import { getSlots } from './slots'
-import { playEmote } from './emotes'
+import { playEmote } from './emote'
 import { applyFacialFeatures, getFacialFeatures } from './face'
 import { setupMappings } from './mappings'
 import { Asset, center, createScene } from './scene'
